feat(sample-react): merge modal samples
