+++ packages/samples/react/src/components/modal/routes.ts
@@ -1,10 +1,8 @@
 import { Routes } from '../../shares/types';
-import { ModalBasic } from './basic';
 import { ModalVariants } from './variants';
 
 export const MODAL_ROUTES: Routes = {
 	modal: {
-		basic: ModalBasic,
 		variants: ModalVariants,
 	},
 };
